Delete alphapy/transforms.py; move date helpers to calendrical
The 76 functions in alphapy/transforms.py were almost entirely OHLCV /
financial technical indicators (ADX, RSI, Bollinger, Heikin-Ashi,
Keltner, Tom DeMark, TTM Squeeze, gaps, VWAP, true_range, Shannon's
demon, etc.) — that surface now lives in alphapy-finance under
mflow/features/transforms.py and mflow/indicators/.

The only callers of transforms in alphapy-pro were the time-series
project's date helpers, so the three generic date functions move to
their natural home in alphapy/calendrical.py:

- dateparts(df, c)  — year/month/day/dayofweek
- timeparts(df, c)  — hour/minute/second
- bizday(df, c)     — date features + business day of month/week

Other changes:

- projects/time-series/config/model.yml: point the transforms section at
  alphapy.calendrical instead of alphapy.transforms.
- alphapy/variables.py: drop the implicit import_module('alphapy.transforms')
  fallback in vfunc; configs must now name the module explicitly (the
  time-series project already does).
- tests/test_imports.py: replace test_transforms_import with
  test_calendrical_import that exercises calendrical.dateparts.
- docs/source/alphapy.rst: drop the alphapy.transforms autodoc section.
